+++ src/xlsxstyles.cpp
@@ -538,8 +538,8 @@ List xlsxstyles::zipFormats(std::vector<xf> styles, bool is_style) {
   colors fonts_color(n);
   colors fills_patternFill_fgColor(n);
   colors fills_patternFill_bgColor(n);
-  colors fills_gradientFill_color1(n);
-  colors fills_gradientFill_color2(n);
+  colors fills_gradientFill_stop1_color(n);
+  colors fills_gradientFill_stop2_color(n);
   colors borders_left_color(n);
   colors borders_right_color(n);
   colors borders_start_color(n); // gnumeric
@@ -562,15 +562,17 @@ List xlsxstyles::zipFormats(std::vector<xf> styles, bool is_style) {
 
   CharacterVector fills_patternFill_patternType(n, NA_STRING);
   CharacterVector fills_gradientFill_type(n, NA_STRING);
-  CharacterVector fills_gradientFill_degree(n, NA_STRING);
-  CharacterVector fills_gradientFill_left(n, NA_STRING);
-  CharacterVector fills_gradientFill_right(n, NA_STRING);
-  CharacterVector fills_gradientFill_top(n, NA_STRING);
-  CharacterVector fills_gradientFill_bottom(n, NA_STRING);
+  IntegerVector fills_gradientFill_degree(n, NA_INTEGER);
+  NumericVector fills_gradientFill_left(n, NA_REAL);
+  NumericVector fills_gradientFill_right(n, NA_REAL);
+  NumericVector fills_gradientFill_top(n, NA_REAL);
+  NumericVector fills_gradientFill_bottom(n, NA_REAL);
+  NumericVector fills_gradientFill_stop1_position(n, NA_REAL);
+  NumericVector fills_gradientFill_stop2_position(n, NA_REAL);
 
-  IntegerVector borders_diagonalDown(n, NA_INTEGER);
-  IntegerVector borders_diagonalUp(n, NA_INTEGER);
-  IntegerVector borders_outline(n, NA_INTEGER);
+  LogicalVector borders_diagonalDown(n, NA_LOGICAL);
+  LogicalVector borders_diagonalUp(n, NA_LOGICAL);
+  LogicalVector borders_outline(n, NA_LOGICAL);
 
   CharacterVector borders_left_style(n, NA_STRING);
   CharacterVector borders_right_style(n, NA_STRING);
@@ -622,8 +624,10 @@ List xlsxstyles::zipFormats(std::vector<xf> styles, bool is_style) {
     fills_gradientFill_right[i] = fill->gradientFill_.right_;
     fills_gradientFill_top[i] = fill->gradientFill_.top_;
     fills_gradientFill_bottom[i] = fill->gradientFill_.bottom_;
-    clone_color(fill->gradientFill_.color1_, fills_gradientFill_color1, i);
-    clone_color(fill->gradientFill_.color2_, fills_gradientFill_color2, i);
+    fills_gradientFill_stop1_position[i] = fill->gradientFill_.stop1_.position_;
+    fills_gradientFill_stop2_position[i] = fill->gradientFill_.stop2_.position_;
+    clone_color(fill->gradientFill_.stop1_.color_, fills_gradientFill_stop1_color, i);
+    clone_color(fill->gradientFill_.stop2_.color_, fills_gradientFill_stop2_color, i);
 
     borders_diagonalDown[i] = border->diagonalDown_;
     borders_diagonalUp[i] = border->diagonalUp_;
@@ -681,6 +685,8 @@ List xlsxstyles::zipFormats(std::vector<xf> styles, bool is_style) {
     fills_gradientFill_right.attr("names") = cellStyles_;
     fills_gradientFill_top.attr("names") = cellStyles_;
     fills_gradientFill_bottom.attr("names") = cellStyles_;
+    fills_gradientFill_stop1_position.attr("names") = cellStyles_;
+    fills_gradientFill_stop2_position.attr("names") = cellStyles_;
     borders_diagonalDown.attr("names") = cellStyles_;
     borders_diagonalUp.attr("names") = cellStyles_;
     borders_outline.attr("names") = cellStyles_;
@@ -726,13 +732,17 @@ List xlsxstyles::zipFormats(std::vector<xf> styles, bool is_style) {
           _["patternType"] = fills_patternFill_patternType),
         _["gradientFill"] = List::create(
           _["type"] = fills_gradientFill_type,
-          _["degree_"] = fills_gradientFill_degree,
+          _["degree"] = fills_gradientFill_degree,
           _["left"] = fills_gradientFill_left,
           _["right"] = fills_gradientFill_right,
           _["top"] = fills_gradientFill_top,
           _["bottom"] = fills_gradientFill_bottom,
-          _["color1"] = list_color(fills_gradientFill_color1, is_style),
-          _["color2"] = list_color(fills_gradientFill_color2, is_style))),
+          _["stop1"] = List::create(
+            _["position"] = fills_gradientFill_stop1_position,
+            _["color"] = list_color(fills_gradientFill_stop1_color, is_style)),
+          _["stop2"] = List::create(
+            _["position"] = fills_gradientFill_stop2_position,
+            _["color"] = list_color(fills_gradientFill_stop2_color, is_style)))),
       _["border"] = List::create(
           _["diagonalDown"] = borders_diagonalDown,
           _["diagonalUp"] = borders_diagonalUp,